+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added JSON as data source, internal data uses a Python code editor, test_data setting is deprecated [#225](https://github.com/golemhq/golem/issues/225)
 - Added 4 test hooks: before_test, before_each, after_each, after_test [#227](https://github.com/golemhq/golem/issues/227)
+- Reworked the report list view: executions are listed mixed and sorted by date desc
 
 ### Deprecated